Amirabdolahian and Ali-Adeeb (2021)

A study on the attitude towards real and computer generated image of people smiling. The response is the latency for a particular electrode located at the back of the brain.

Usage

AA21

Format

A data frame with 1169 rows and 3 variables:

id

[factor] participant identifier

stimulus

[factor] image shown, one of real image (real), a one-pass generative adversarial network generate image with slightly happy face (GAN1) and a two-pass GAN with an extremely happy person (GAN2)

epoch

[factor] ordering of picture

latency

[double] latency of N170 EEG signal for Fz electrode

Source

Saeed Amirabdolahian, personal communication, licensed under CC BY-NC-SA 4.0

Details

The data were collected as part of a course project during Winter 2021 (TECH 80747A Information Technologies and Neuroscience), which took place during the Covid19 pandemic. Each student participant had his own apparatus, which explains the large between-participant measurement error.

References

S. Amirabdolahian and R. Ali-Adeeb (2021). Measuring face images visual fidelity in neurophysiological level: a study on facial expression manipulating technologies, technical report, HEC Montreal.