Brucks and Levav (2022), Lab experiment

Brucks and Levav (2022), Lab experiment

Usage

BL22_L

Format

A data frame with 301 rows and 8 variables:

cond

[factor] experimental condition, either video or f2f for face-to-face

nidea

[integer] number of ideas generated by pair

ncreative

[integer] number of ideas that were above the average creativity score

select_creative

[double] creativity score of the idea each group selected as their most creative idea

top_creative

[double] creativity score of the highest scoring idea in each group

avg_creative

[double] average creativity of all ideas generated by each group

study

[factor] study number

Source

Research Box 282, https://researchbox.org/282, licensed under CC BY 4.0

Details

The authors recruted 301 pairs of people and assigned them to either face-to-face interactions or videocall. They measured the number of ideas generated during the discussion as well as the fraction of those which were deemed creative using a `creativity_score`. The latter was measured by two undergraduates who evaluated all ideas based on novelty using a Likert scale with 7 points.

References

Brucks, M.S. and J. Levav (2022). Virtual communication curbs creative idea generation. Nature 605, 108–112. doi:10.1038/s41586-022-04643-y