Bertrand and Mullainathan (2004), Table 2

Bertrand and Mullainathan (2004), Table 2

Usage

BM04_T2

Format

a data frame in long format of 9 observations on the following variables

black

number of black partipants called back, out of two

white

number of white partipants called back, out of two

count

[integer] count of occurence of each combination

References

Bertrand, M., and S. Mullainathan. 2004. Are Emily and Greg More Employable Than Lakisha and Jamal? A Field Experiment on Labor Market Discrimination. American Economic Review 94 (4): 991-1013. doi:10.1257/0002828042002561 .