Bobak, Mileva and Hancock (2019), Experiment 2

Data on matching of pairs of pictures from official ID depending on the color. Each of the 52 participants was assigned pictures of celebrities and needed to decide whether pairs of picture (potentially mixed or both monochrome or color) were of the same person.

Usage

BMH19_E2

Format

A data frame with 156 rows and 6 variables:

id

[integer] participant ID

gender

[factor] gender of participant, either male or female

age

[integer] age of participant

color

[factor] color of photographs, either grayscale (mon), colored (col) or mixed (mix)

accuracy

[double] percentage of accuracy (same)

pcorr

[double] percentage of correctly detected match (different)

Source

Supplementary material, article licensed under CC BY 4.0

References

Bobak, A. K., Mileva, V. R., & Hancock, P. J. B. (2019). A grey area: How does image hue affect unfamiliar face matching? Cognitive Research: Principles and Implications, 4(1), 27. doi:10.1186/s41235-019-0174-3