Brodeur et al. (2021), exploratory data analysis

Experiment from Tech3Lab using a driving simulator to study the impact of various distraction on attention and performance. The experimental conditions were compared in Table 3 of Brodeur et al. (2021) using Wilcoxon's signed rank test.

Usage

BRLS21_T3

Format

A data frame with 124 rows and 3 variables:

id

[factor] participant identifier

task

[factor] type of distraction

nviolation

[integer] total number of violations

Source

Shang Lin Chen, personal communication, distributed under CC BY-NC-SA 4.0

References

Brodeur, M., Ruer, P. Léger, P. and S. Sénécal (2021). Smartwatches are more distracting than mobile phones while driving: Results from an experimental study, Accident Analysis & Prevention, 149, 1-9. doi:10.1016/j.aap.2020.105846