Baumann, Seifert-Kessell and Jones (1992)

Experimental study on the effectiveness of different reading strategies on understanding of children.

Usage

BSJ92

Format

A data frame with 66 rows and 6 variables:

group

[factor] experimental group, one of directed reading-thinking activity (DRTA), think-aloud (TA) and directed reading group (DR)

pretest1

[integer] score (out of 16) on pretest for the error detection task

pretest2

[integer] score (out of 15) on second pretest using a comprehension monitoring questionnaire

posttest1

[integer] score (out of 16) on an error detection task

posttest2

[integer] score (out of 18) on an expanded comprehension monitoring questionnaire

posttest3

[integer] number of correct item, ranging from 0 to 56 on the Degrees of Reading Power test, a cloze test

Source

Obtained from `carData` package (`Baumann`)

References

Baumann, J. F., Seifert-Kessell, N., & Jones, L. A. (1992). Effect of Think-Aloud Instruction on Elementary Students’ Comprehension Monitoring Abilities. Journal of Reading Behavior, 24(2), 143–172. doi:10.1080/10862969209547770