Chandler (2016) replication

Replication by Chandler (2016) of Study 4a of Janiszewski and Uy (2008). Both studies measured the amount of adjustment when presented with vague or precise range of value for objects, with potential encouragement for adjusting more the value.

Usage

C16

Format

A data frame with 120 rows and 4 variables:

anchor

[factor] anchor, either round or precise

magnitude

[factor] magnitude of anchor, 0 or 1

gender

[factor] participant gender, either female, male or other

madjust

[double] mean adjustement

Source

Open Science Foundation, https://osf.io/aaudl, licensed under CC0 1.0 Universal

References

Chandler, J. J. (2016). Replication of Janiszewski & Uy (2008, Psychological Science, study 4b). OSF. https://osf.io/aaudl

Janiszewski, C., & Uy, D. (2008). Precision of the anchor influences the amount of adjustment. Psychological Science, 19(2), 121–127. doi:10.1111/j.1467-9280.2008.02057.x