Curley et al. (2022)

This study aimed at finding the impact of having multiple options for jurors as is the case in Scotland (guilty, not guilty, not proven), relative to the usual setting of guilty or not guilty verdict.

The response was measured using the sum of the scores for 11 ratings (rounds) on a Likert scale ranging from 0 (no evidence of guilt) to 14 (guilty).

Usage

C22

Format

A data frame with 256 rows and 7 variables:

id

[factor] participant identifier

anchor

[factor] order of evidence, (strong-first versus weak-first)

vignette

[factor] homocide vignette, either 1 or 2

vorder

[factor] which vignette was presented first?

verdictsyst

[factor] verdict system, either two or three possible outcomes.

guilt

[double] final belief of guilt score, with lower values representing no belief of guilt

pjaq

[integer] PJAQ score, a 29-item questionnaire, which measures total pre-trial bias

Source

Open Science Foundation, https://osf.io/zg9hw/, unspecified license

References

Lee J. Curley, Jennifer Murray, Rory MacLean, James Munro, Martin Lages, Lara A. Frumkin, Phyllis Laybourn & David Brown (2022) Verdict spotting: investigating the effects of juror bias, evidence anchors and verdict system in jurors, Psychiatry, Psychology and Law, 29(3), 323-344, doi:10.1080/13218719.2021.1904450