Duke and Amir (2023), Experiment 1

This is a subset of the data from Experiment 1 of Duke and Amir (2022). The study investigated the impact of presenting customers with a sequential choice (first decide whether or not to buy, then pick quantity) as opposed to an integrated decision (choose not to buy, or one of different quantities) on sales and the number of items bought.

Usage

DA23_E1

Format

A data frame with 397 rows and 5 variables:

format

[factor] experimental condition, either quantity-integrated or quantity-sequential

purchased

[integer] binary variable indicating whether the respondant purchased the item (1) or not (0)

amount

[integer] amount purchased

gender

[factor] gender of respondant, one of female, male or other

age

[integer] age of respondant

Source

Research Box 602, https://researchbox.org/602, licensed under CC BY 4.0

References

Duke, K.E. and O. Amir (2023). The Importance of Selling Formats: When Integrating Purchase and Quantity Decisions Increases Sales, Marketing Science,42(1), 87-109. doi:10.1287/mksc.2022.1364