Duke and Amir (2023), Experiment 3

This is a data from Experiment 3 of Duke and Amir (2022), who studied the impact of sequential vs integrated quantity purchase. The authors manipulated the company name (same ratings, same product and price) and offered participants the option to purchase up to two items on sale from either, with brand order counterbalanced. The study investigated the impact of presenting customers with a sequential choice (first decide whether or not to buy, then pick quantity) as opposed to an integrated decision (choose not to buy, or one of different quantities) on sales and the number of items bought.

Usage

DA23_E3

Format

A data frame with 399 rows and 6 variables:

format

[factor] experimental condition, either quantity-integrated or quantity-sequential

purchased

[integer] binary variable indicating whether the respondant purchased the item (1) or not (0)

amount

[integer] amount of items purchased

firstbrand

[factor] company name offering the product, counterbalanced

gender

[factor] gender of respondant, one of female, male or other

age

[integer] age of respondant

Source

Research Box 602, https://researchbox.org/602, licensed under CC BY 4.0

References

Duke, K.E. and O. Amir (2023). The Importance of Selling Formats: When Integrating Purchase and Quantity Decisions Increases Sales, Marketing Science,42(1), 87-109. doi:10.1287/mksc.2022.1364

Examples

mod <- glm(purchased ~ format * firstbrand, data = DA23_E3)
summary(mod)
#> 
#> Call:
#> glm(formula = purchased ~ format * firstbrand, data = DA23_E3)
#> 
#> Coefficients:
#>                                                     Estimate Std. Error t value
#> (Intercept)                                          0.75000    0.04599  16.307
#> formatquantity-sequential                           -0.07000    0.06504  -1.076
#> firstbrandfor your beauty                           -0.01733    0.06488  -0.267
#> formatquantity-sequential:firstbrandfor your beauty -0.05043    0.09211  -0.547
#>                                                     Pr(>|t|)    
#> (Intercept)                                           <2e-16 ***
#> formatquantity-sequential                              0.282    
#> firstbrandfor your beauty                              0.790    
#> formatquantity-sequential:firstbrandfor your beauty    0.584    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for gaussian family taken to be 0.2115379)
#> 
#>     Null deviance: 84.697  on 398  degrees of freedom
#> Residual deviance: 83.557  on 395  degrees of freedom
#> AIC: 518.51
#> 
#> Number of Fisher Scoring iterations: 2
#>