Dorison and Minson (2022), Study 1

Participants (American citizens) in the study were paired with people from the opposing party (Republican or Democrat) and asked to perform a 10 minutes presentation to discuss beliefs about the best candidate, with pointers to subjects such as Covid19, environmental policy and immigration. Only people who interacted via the chat are kept.

Usage

DM22_S1

DM22_S1_long

Format

A data frame with 367 rows and 17 variables:

A data frame with 734 rows and 4 variables:

anxiety

[numeric] anxiety score for average of four nine point Likert scales

target

[factor] categorical variable, self or other

vote

Who are you most likely to vote for in the upcoming U.S. presidential election (2020)? either Republican candidate (1) or Democratic candidate (2)

group

[factor] group label

Source

ResearchBox 577, https://researchbox.org/577, licensed under CC BY 4.0

Functions

DM22_S1: Wide-format with paired observations

age

[integer] age of participant

gender

[factor] gender of participant, either male or female

pos_emotion_part

[integer] partner positive emotion index

neg_emotion_part

[integer] partner negative emotion index

wordcount_part

[integer] partner total words in debate

hedges_part

[integer] number of hedges for partners

pos_emotion

[integer] positive emotion index

neg_emotion

[integer] negative emotion index

wordcount

[integer] total words in debate

hedges

[integer] number of hedges

group

[factor] group label

anxiety_self

[numeric] anxiety score for self, average of four nine point Likert scales

anxiety_other

[numeric] anxiety score for other, average of four nine point Likert scales

vote

Who are you most likely to vote for in the upcoming U.S. presidential election (2020)? either Republican candidate (1) or Democratic candidate (2)

polideo

self-reported political orientation, with choices very/somewhat/slightly liberal (1/2/3), neither (4), to slightly/somewhat/very conservative (5/6/7).

DM22_S1_long: Long-format database

References

Dorison CA, Minson JA. (2022). You can’t handle the truth! Conflict counterparts over-estimate each other’s feelings of self-threat. Organizational Behavior and Human Decision Processes, 170, 104147. doi:10.1016/j.obhdp.2022.104147