Hart and Schweitzer (2022) pilot study

The authors measured the economic relevance of relational outcomes (ERRO), by substracting the participants rating for good versus bad relationships, which rated from -10 (much worse) to 10 (much better). The data are presented in long format for each of the three items/services split by category.

Usage

HS22_P

Format

A data frame with 5808 rows and 7 variables:

id

[factor] participant identifier

score

[double] ERRO score

item

[factor] item

category

[factor] category of the item, either product or service

negotcl

[factor] took negotiation class

gender

[factor] gender of participant, either female or male

age

[integer] age of the participant

Source

Research Box 366, https://researchbox.org/366, licensed under CC BY 4.0

References

Hart E, Schweitzer ME. (2022) When we should care more about relationships than favorable deal terms in negotiation: The economic relevance of relational outcomes (ERRO)'. Organizational Behavior and Human Decision Processes. 168:104108. doi:10.1016/j.obhdp.2021.104108

Examples

# Repeated measure ANOVA
rmod <- afex::aov_ez(
    id = "id",
    dv = "score",
    within = "item",
    data = ERRO_pilot_long)
#> Error: object 'ERRO_pilot_long' not found
 emmeans::emmeans(object = rmod, specs = "item") |>
   emmeans::contrast(list(service_vs_product = c(1, -1, 1, -1, 1, -1)/3))
#> Error: object 'rmod' not found