Hart and Schweitzer (2022), Study 2

Between-subject mediation study. The authors investigated the trade-offs between relational and deal terms concerns. Comparisons are made between high ERRO (for service purchase from someone providing his services as a house cleaner, versus from a seller of a vacuum cleaner). Participants were asked to allocate 4 points (for choice 1) to either deal or relationship and seven points for choice 2. The slope variable is obtained `((7-dealpts2) - (4-dealpts1))/(dealpts2 - dealpts1)`

Usage

HS22_S2

Format

A data frame with 295 rows and 6 variables:

erro

[factor] whether the offer is for a service (cleaning), or for a product (selling a vacuum cleaner)

relation

[double] average relationship importance

deal

[double] average deal importance

slope

[double] linear slope between choices 1 and 2

dealpts1

[double] number of points out of four allocated to deal terms for choice 1. The balance is allocated to relation.

dealpts2

[double] number of points out of seven allocated to deal terms for choice 2. The balance is allocated to relation.

Source

Research Box 366, https://researchbox.org/366, licensed under CC BY 4.0

References

Hart E, Schweitzer ME. (2022) When we should care more about relationships than favorable deal terms in negotiation: The economic relevance of relational outcomes (ERRO)'. Organizational Behavior and Human Decision Processes. 168:104108. doi:10.1016/j.obhdp.2021.104108