Johnson, Cheung and Donnellan (2014), Study 2

This dataset is for a replication of a study of Schnall, Benton, and Harvey (2008), who hypothesied that cleanliness (washing hands) reduces the severity of moral judgments. The study failed to replicate.

Usage

JCD14_S2

Format

A data frame with 126 rows and 20 variables:

condition

[factor] experimental condition, either control or cleanliness

mvignette

[double] mean of vignettes

dog

[double] Likert scale for the dog vignette

trolley

[double] Likert scale for the trolley vignette

wallet

[double] Likert scale for the wallet vignette

plane

[double] Likert scale for the plane vignette

resume

[double] Likert scale for the resume vignette

kitten

[double] Likert scale for the kitten vignette

age

[integer] age of participant

gender

[factor] gender of participant, either male or female

Source

Open Science Foundation, https://osf.io/zwrxc, unspecified license

Details

The vignettes (dog, trolley, wallet, plane, resume, kitten) are measured using a Likert scale ranging from nothing wrong at all (1) to extremely wrong (7).

References

Johnson, D.J., F. Cheung and M.B. Donnellan (2014), Does Cleanliness Influence Moral Judgments?, Social Psychology, 45(3), 209-215, doi:10.1027/1864-9335/a000186