Koetke et al. (2024), Study 4

Two (humility) by three race) between subject design to study participant perception of scientists depending on the intellectual humility. The authors performed linear mediation analysis of METI by ih, research by METI and by ih, separately. They also considered two-way ANOVA models with controls for competence and warmth as control variates.

Usage

KSBN24_S4

Format

A data frame with 371 rows and 10 variables:

METI

[double] METI scale for the trust in the scientist

research

[double] support for Dr. Wilson's research

ih

[double] perceived intellectual humility score

competence

[double] score for the question "As viewed by society, how competent is Dr."

warmth

[double] score for the question "As viewed by society, how warm is Dr."

condition

[factor] experimental factor for intellectual humility, either low, high or neutral

race

[factor] experimental factor for race, one of latinx, black (African American) or white

politideo

[factor] political ideology on Prolifics platform, either liberal or conservative

age

[integer] age of participant

gender

[factor] gender of participant, either male, female or other

Source

Open Science Foundation, https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961, unspecified license

References

Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. Nature Human Behaviour (2024). doi:10.1038/s41562-024-02060-x . Preprint at https://osf.io/preprints/psyarxiv/ah3fp/