Koetke et al. (2024), Study 5

Authors perform one-way ANOVA for CIHS score and the subscales as a function of condition, and the same for the METI scale and subscales. Three independent linear mediation are tested for research average and METI, with the experimental condition mediated by the intellectual humility score ih, and research average by condition also mediated by METI.

Usage

KSBN24_S5

Format

A data frame with 679 rows and 13 variables:

ih

[double] CIHS score for intellectual humility of scientist

condition

[factor] experimental condition for intellectual humility, either low or high

indep

[double] CIHS subscale score for independent of science

respect

[double] CIHS subscale score for respect of science

openness

[double] CIHS subscale score for openness of science

lackover

[double] CIHS subscale score for lack over

METI

[double] METI scale for the trust in science

expertise

[double] METI subscale for expertise

integrity

[double] METI subscale for integrity

benevolence

[double] METI subscale for benevolence

research

[double] score for support for the research

warmth

[double] score for warmth

competence

[double] score for competence

Source

Open Science Foundation, https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961, unspecified license

References

Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. Nature Human Behaviour (2024). doi:10.1038/s41562-024-02060-x . Preprint at https://osf.io/preprints/psyarxiv/ah3fp/