Koetke et al. (2024), Study 5 (long format)

Data from Study 5 in long format for the questions related to trust. The authors performed repeated measure ANOVA for each of the three response in term. Also suitable for multivariate analysis of variance. All responses are measured using a Likert scale ranging from 1 (strongly disagree) to 7 (strongly agree).

Usage

KSBN24_S5L

Format

A data frame with 4074 rows and 5 variables:

id

[factor] participant identifier

statement

[factor] statement of the question

pih

[double] perceived intellectual humility for the statement

trustdr

[double] score for trust in doctor Wilson

trustresearch

[double] score for trust in research results of doctor Wilson

Source

Open Science Foundation, https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961, unspecified license

References

Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. Nature Human Behaviour (2024). doi:10.1038/s41562-024-02060-x . Preprint at https://osf.io/preprints/psyarxiv/ah3fp/