Lee and Choi (2019), Table 2

Contingency table for Study 3, measuring the expected number of products as a function of the text description and image showing either one or six toothbrushes.

Usage

LC19_T2

Format

A data frame with 12 rows and 4 variables:

text

[factor] number of toothbrush in text description

image

[factor] number of toothbrushes displayed in image

expected

[factor] expected number of toothbrushes

count

[integer] count of occurence

Source

Mendeley, doi:10.17632/r9t7hh7cy3.1 , licensed under CC BY 4.0

References

K. Lee and J. Choi (2019). Image-text inconsistency effect on product evaluation in online retailing, Journal of Retailing and Consumer Services, 49, 279-288, doi:10.1016/j.jretconser.2019.03.015