ManyLab2 replication of Risen and Gilovich (2008)

The data is a replication with undergraduate students of a study. The authors measured the likelihood of bad outcome when tempting fate by not preparing in advance for a task. The replication data are unbalanced and the study failed to replicate

Usage

MANY18_S18

Format

A data frame with 4433 rows and 4 variables:

lab

[factor] laboratory identifier

likelihood

[integer] likelihood of bad outcome, from 1 to 10

gender

[factor] gender of participant, either male or female

condition

[factor] experimental condition, either prepared or unprepared

Source

Open Science Foundation, https://osf.io/sg3su, licensed under CC0 1.0 Universal

References

Klein, R. A., et al. (2018). Many labs 2: Investigating variation in replicability across samples and settings. Advances in Methods and Practices in Psychological Science, 1(4), 443–490. doi:10.1177/2515245918810225 Risen, J. L., and T. Gilovich, (2008). Why people are reluctant to tempt fate. Journal of Personality and Social Psychology, 95(2), 293–307. doi:10.1037/0022-3514.95.2.293