Morgan and Halfaker (2018), Table 1

Data on the impact of Wikipedia Teahouse on the number of edits. The authors looked at barriers to participation

Usage

MH18_T1

Format

A data frame with 12 rows and 4 variables:

neweditors

[integer] number of new editors

period

[factor] number of weeks since joining Wikipedia

group

[factor] experimental condition, either invited to teahouse or control

edits

[factor] level of activity, either 1-4 edits or 5+ edits

References

Jonathan T. Morgan and Aaron Halfaker. 2018. Evaluating the impact of the Wikipedia Teahouse on newcomer socialization and retention. In Proceedings of the 14th International Symposium on Open Collaboration (OpenSym '18). Association for Computing Machinery, New York, NY, USA, 20, 1–7. doi:10.1145/3233391.3233544