Elliot et al. (2021), dataset 1

Data for Table 2 of Elliot et al. (2021), split by replicating lab. The data are counts of verbalization, suitable for a generalized linear mixed model and chi-square test for contingency data.

Usage

MULTI21_D1

Format

A data frame with 204 rows and 4 variables:

lab

[factor] location of laboratory

age

[factor] age group, one of 5yo, 6yo, 7yo or 10yo

frequency

[factor] how often students verbalized during task, one of never, sometimes or usually

count

[integer] count of the number of children

Source

Open Science Foundation, https://osf.io/vgxkf, licensed under CC-BY Attribution 4.0 International

References

Elliott, E. M., C. C. Morey, A.M. AuBuchon et al. (2021). Multilab Direct Replication of Flavell, Beach, and Chinsky (1966): Spontaneous Verbal Rehearsal in a Memory Task as a Function of Age. Advances in Methods and Practices in Psychological Science. doi:10.1177/25152459211018187