Elliot et al. (2021), dataset 2

Data for a linear mixed model and a two-way analysis of variance with age and timing.

Usage

MULTI21_D2

Format

A data frame with 2931 rows and 7 variables:

id

[factor] unique participant ID

lab

[factor] location of laboratory

taskorder

[factor] order of tasks for individual, either immediate first or delayed first

timing

[factor] timing of the task for recall, either delayed or immediate or point-and-name

age

[factor] age group, one of 5yo, 6yo, 7yo or 10yo

mcorrect

[integer] number of words correctly recalled, from 1 to 5

frequency

[factor] how often students verbalized during task, one of never, sometimes or usually

Source

Open Science Foundation, https://osf.io/vgxkf, licensed under CC-BY Attribution 4.0 International

References

Elliott, E. M., C. C. Morey, A.M. AuBuchon et al. (2021). Multilab Direct Replication of Flavell, Beach, and Chinsky (1966): Spontaneous Verbal Rehearsal in a Memory Task as a Function of Age. Advances in Methods and Practices in Psychological Science. doi:10.1177/25152459211018187