Moon and vanEpps (2022+), Study 1

Study 1 considers proportion of donators to a charity. Participants in the online panel were provided with an opportunity to win 25$ and donate part of this amount to a charity of their choosing. The data provided include only people who did not exceed this amount and indicated donating a non-zero amount.

Usage

MV23_S1

Format

A data frame with 869 rows and 4 variables:

before

[integer] did people donate before to this charity? 0 for no, 1 for yes.

donate

[integer] did people choose to donate to the charity? 0 for no, 1 for yes.

condition

[factor] experimental condition, either an open-ended amount or a suggested quantity

amount

[double] amount of proposed donation, NA if the person declined to donate

Source

Research Box 54, https://researchbox.org/54, licensed under CC BY 4.0

References

Moon, A. and EM VanEpps (2023). Giving Suggestions: Using Quantity Requests to Increase Donations, Journal of Consumer Research, 50(1), 190–210. doi:10.1093/jcr/ucac047

Examples

ctab <- with(MV23_S1, table(donate, condition))
# Chi-square test of independence
chisq_test <- chisq.test(ctab, correct = FALSE)
chisq_test
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  ctab
#> X-squared = 20.993, df = 1, p-value = 4.61e-06
#> 
# Effect size
sqrt(chisq_test$statistic/nrow(MV23_S1))
#> X-squared 
#> 0.1554272 
# Welch t-test
t.test(amount ~ condition, MV23_S1)
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  amount by condition
#> t = -1.1181, df = 581.55, p-value = 0.264
#> alternative hypothesis: true difference in means between group open-ended and group quantity is not equal to 0
#> 95 percent confidence interval:
#>  -1.7398139  0.4775039
#> sample estimates:
#> mean in group open-ended   mean in group quantity 
#>                 10.32116                 10.95232 
#>