Moon and vanEpps (2023), Study 4

The authors studied the impact on donation amount of manipulating the minimal suggested amount. The benchmark was an open-ended box, with quantity featuring a minimum value, 15$, 25$ or an alternative amount. The authors also recorded whether the person had donated before.

Usage

MV23_S4

Format

A data frame with 996 rows and 3 variables:

donate

[integer] did people choose to donate (donate before to this charity? 0 for no, 1 for yes.

condition

[factor] experimental condition, either an open-ended amount or a suggested amount of quantity1, quantity5 or quantity10 for one, five and 10 dollars.

amount

[double] amount of proposed donation, NA if the person declined to donate

Source

Research Box 54, https://researchbox.org/54, licensed under CC BY 4.0

References

Moon, A. and EM VanEpps (2023). Giving Suggestions: Using Quantity Requests to Increase Donations, Journal of Consumer Research, 50(1), 190–210.doi:10.1093/jcr/ucac047