Roczniewska and Higgins (2019), Study 1

Study of mediator following changes in the Polish education system, whereby middle school was due to disappear and be replaced with a new system with 8 years of elementary school and four years of high school. The authors surveyed teachers in middle and primary schools with two messages, either prevention or promotion.

Usage

RH19_S1

Format

A data frame with 198 rows and 12 variables:

change

[factor] change outcome depending on school type, either middle school (negative) or primary school (positive)

regfocus

[factor] situational regulatory focus manipulation: education change message framing, either prevention or promotion message framing

fluency

[double] mean for fluency items

engagement

[double] mean for engagement items

fairness

[double] mean for change process fairness perceptions items

dopenness

[double] mean for change openness items

promem

[double] mean for work regulatory focus promotion items

prevem

[double] mean for work regulatory focus prevention items

gender

[factor] gender of teacher, either man or woman

age

[integer] age of teacher

tenure

[integer] number of years teaching

hours_week

[integer] weekly number of hours working

Source

Open Science Foundation, https://osf.io/vrazp/, unspecified license

References

M. Roczniewska and Higgins, E.T. (2019). Messaging organizational change: How regulatory fit relates to openness to change through fairness perceptions, Journal of Experimental Social Psychology, 85, doi:10.1016/j.jesp.2019.103882