de la Rosa and Tully (2022), Study 2

Participants played a life simulator game with different amounts transferred to their "bank account", either on a daily or weekly base. They had to make 15 binary spending decisions. Excess expenses resulted in overdraft fees of 35 dollars. The authors recorded the total amount spent, the sequential decisions and the number of overdraft charges.

Usage

RT22_S2

Format

A data frame with 401 rows and 8 variables:

gender

[factor] gender of participant

age

[numeric] age

decision

[integer] To what extent did the praticipant try to make decisions as they would in real life

nexpdec

[integer] number of expensive decisions

totspend

[numeric] total spending

nover

[integer] number of overdraft charges (35 each, in addition to total spending)

condition

[factor] experimental condition, either week for weekly pay or day for daily transfer

subwealth

[numeric] subjective wealth index, ranging from low (1) to high (7)

Source

ResearchBox 231, https://researchbox.org/231, licensed under CC BY 4.0

References

Wendy De La Rosa, Stephanie M Tully (2022). The Impact of Payment Frequency on Consumer Spending and Subjective Wealth Perceptions, Journal of Consumer Research, 48(6), 991-1009, doi:10.1093/jcr/ucab052