de la Rosa and Tully (2022), Study 5

Study 5 is a replication of Study 2, with additionalmeasures of intertemporal discount rate, financial literacy, Participants played a four-week life simulator and received either 1400 biweekly (starting on the first Friday), or 140 daily on weekdays. This ensured different balances even with the same spending decisions. Participants had to choose 28 daily decisions. There was no overdraft fee in this study.

gender

[factor] gender of participant

age

[integer] age

subjectivewealth

[numeric] average of six Likert scales on subjective wealth, on a scale of 0 (low) to 100 (high)

preduncertain

[numeric] average of four Likert scales on prediction uncertainty, on a scale of 0 to 100

valence

[integer] reported valence from the 3-item self-assessment Manikin scale

arousal

[integer] reported arousal from the 3-item self-assessment Manikin scale

power

[integer] reported power from the 3-item self-assessment Manikin scale

condition

[factor] experimental condition, either bi-weekly for biweekly pay or dayly for daily transfer

nexpdec

[integer] number of expensive decisions

totspend

[numeric] total spending

Usage

RT22_S5

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1120 rows and 10 columns.

Source

ResearchBox 231, https://researchbox.org/231, licensed under CC BY 4.0

References

Wendy De La Rosa, Stephanie M Tully (2022). The Impact of Payment Frequency on Consumer Spending and Subjective Wealth Perceptions, Journal of Consumer Research, 48(6), 991-1009, doi:10.1093/jcr/ucab052